refactor: rearrange files
